requirements.txt
Flask
mysql-connector-python
pillow
